requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
azure-core==1.31.0
azure-cosmos==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.3
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
idna==3.10
iniconfig==2.0.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-mock==3.14.0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
